test(api): share event service fixtures
