+++ crates/api/src/event_service.rs
@@ -630,420 +630,21 @@ impl proto::event_service_server::EventService for EventService {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::peer_types::{
-        POLICY_EVENT_HISTORY_CAPACITY, PolicyEvent, SESSION_EVENT_HISTORY_CAPACITY, SessionEvent,
-        SessionLifecycleEvent, SessionLifecycleEventType, SessionNotificationEvent,
-        SessionNotificationEventType,
-    };
+    use crate::peer_types::{PolicyEvent, SessionLifecycleEventType, SessionNotificationEventType};
     use crate::proto::event_service_server::EventService as EventServiceTrait;
-    use crate::test_support::metrics_text as gather_text;
-    use rustbgpd_fsm::SessionState;
-    use rustbgpd_rib::{EvpnRouteEvent, RouteEvent};
-    use rustbgpd_wire::{
-        EthernetSegmentIdentifier, EthernetTagId, EvpnMacIp, EvpnRoute, MacAddress, MplsLabel,
-        Origin, PathAttribute, RouteDistinguisher,
+    use crate::test_support::{
+        blackhole_status, dataplane_route_event, evpn_event, fib_status, lifecycle_event,
+        metrics_text as gather_text, notification_event, policy_event, route_event, session_event,
+        spawn_fake_evpn_rib, spawn_fake_peer_manager, spawn_fake_peer_manager_with_history,
+        spawn_fake_peer_manager_with_policy_history, spawn_fake_rib,
+        spawn_fake_rib_with_evpn_history,
     };
+    use rustbgpd_rib::RouteEvent;
     use rustbgpd_wire::{Ipv4Prefix, Ipv6Prefix};
     use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
     use std::sync::{Arc, Mutex};
-    use tokio::sync::{broadcast, mpsc};
+    use tokio::sync::broadcast;
     use tokio_stream::StreamExt;
-
-    fn route_event(prefix: Prefix, peer: IpAddr) -> RouteEvent {
-        RouteEvent {
-            event_id: 1,
-            event_type: RouteEventType::Added,
-            prefix,
-            peer: Some(peer),
-            previous_peer: None,
-            target_peer: None,
-            timestamp: "123".to_string(),
-            path_id: 7,
-            reason: String::new(),
-        }
-    }
-
-    fn evpn_route(peer: IpAddr, mac_byte: u8) -> rustbgpd_rib::route::EvpnRibRoute {
-        let route = EvpnRoute::MacIp(EvpnMacIp {
-            rd: RouteDistinguisher::new([0, 0, 0xFD, 0xE8, 0, 0, 0, 100]),
-            esi: EthernetSegmentIdentifier::ZERO,
-            ethernet_tag: EthernetTagId(0),
-            mac: MacAddress::new([mac_byte; 6]),
-            ip: Some(IpAddr::V4(Ipv4Addr::new(10, 0, 0, mac_byte))),
-            label1: MplsLabel::new(100),
-            label2: None,
-        });
-        rustbgpd_rib::route::EvpnRibRoute {
-            route,
-            next_hop: peer,
-            link_local_next_hop: None,
-            peer,
-            attributes: Arc::new(vec![PathAttribute::Origin(Origin::Igp)]),
-            received_at: std::time::Instant::now(),
-            origin_type: rustbgpd_rib::route::RouteOrigin::Ibgp,
-            peer_router_id: Ipv4Addr::new(192, 0, 2, mac_byte),
-            is_stale: false,
-            is_llgr_stale: false,
-        }
-    }
-
-    fn evpn_event(
-        event_type: RouteEventType,
-        peer: Option<IpAddr>,
-        previous_peer: Option<IpAddr>,
-    ) -> EvpnRouteEvent {
-        let route = peer.map(|peer| evpn_route(peer, 1));
-        let previous_best = previous_peer.map(|peer| evpn_route(peer, 2));
-        let key = route
-            .as_ref()
-            .or(previous_best.as_ref())
-            .expect("test event needs a route")
-            .key();
-        EvpnRouteEvent {
-            event_type,
-            key,
-            best: route,
-            previous_best,
-            peer,
-            previous_peer,
-            timestamp: "123".to_string(),
-        }
-    }
-
-    fn spawn_fake_rib() -> (mpsc::Sender<RibUpdate>, broadcast::Sender<RouteEvent>) {
-        let (rib_tx, mut rib_rx) = mpsc::channel(16);
-        let (events_tx, _) = broadcast::channel(16);
-        let (evpn_events_tx, _) = broadcast::channel(16);
-        let events_tx_for_task = events_tx.clone();
-        let evpn_events_tx_for_task = evpn_events_tx.clone();
-        tokio::spawn(async move {
-            while let Some(update) = rib_rx.recv().await {
-                match update {
-                    RibUpdate::SubscribeRouteEvents { reply } => {
-                        let _ = reply.send(events_tx_for_task.subscribe());
-                    }
-                    RibUpdate::SubscribeEvpnRouteEvents { reply } => {
-                        let _ = reply.send(evpn_events_tx_for_task.subscribe());
-                    }
-                    RibUpdate::QueryEvpnRouteEventHistory { reply, .. } => {
-                        let _ = reply.send(Vec::new());
-                    }
-                    _ => {}
-                }
-            }
-        });
-        (rib_tx, events_tx)
-    }
-
-    fn spawn_fake_rib_with_evpn_history(history: Vec<EvpnRouteEvent>) -> mpsc::Sender<RibUpdate> {
-        let (rib_tx, mut rib_rx) = mpsc::channel(16);
-        let (events_tx, _) = broadcast::channel(16);
-        let history = Arc::new(history);
-        tokio::spawn(async move {
-            while let Some(update) = rib_rx.recv().await {
-                match update {
-                    RibUpdate::SubscribeRouteEvents { reply } => {
-                        let _ = reply.send(events_tx.subscribe());
-                    }
-                    RibUpdate::QueryEvpnRouteEventHistory {
-                        peer,
-                        route_type,
-                        rd,
-                        event_types,
-                        limit,
-                        reply,
-                    } => {
-                        let limit = if limit == 0 { 4096 } else { limit.min(4096) };
-                        let mut events: Vec<_> = history
-                            .iter()
-                            .rev()
-                            .filter(|event| {
-                                route_type
-                                    .is_none_or(|route_type| event.key.route_type() == route_type)
-                            })
-                            .filter(|event| {
-                                rd.is_none_or(|rd| {
-                                    rustbgpd_rib::event::evpn_key_rd(&event.key) == rd
-                                })
-                            })
-                            .filter(|event| {
-                                event_types.is_empty() || event_types.contains(&event.event_type)
-                            })
-                            .filter(|event| match peer {
-                                Some(peer) => {
-                                    event.peer == Some(peer) || event.previous_peer == Some(peer)
-                                }
-                                None => true,
-                            })
-                            .take(limit)
-                            .cloned()
-                            .collect();
-                        events.reverse();
-                        let _ = reply.send(events);
-                    }
-                    _ => {}
-                }
-            }
-        });
-        rib_tx
-    }
-
-    fn spawn_fake_evpn_rib() -> (mpsc::Sender<RibUpdate>, broadcast::Sender<EvpnRouteEvent>) {
-        let (rib_tx, mut rib_rx) = mpsc::channel(16);
-        let (route_events_tx, _) = broadcast::channel(16);
-        let (evpn_events_tx, _) = broadcast::channel(16);
-        let evpn_events_tx_for_task = evpn_events_tx.clone();
-        tokio::spawn(async move {
-            while let Some(update) = rib_rx.recv().await {
-                match update {
-                    RibUpdate::SubscribeRouteEvents { reply } => {
-                        let _ = reply.send(route_events_tx.subscribe());
-                    }
-                    RibUpdate::SubscribeEvpnRouteEvents { reply } => {
-                        let _ = reply.send(evpn_events_tx_for_task.subscribe());
-                    }
-                    RibUpdate::QueryEvpnRouteEventHistory { reply, .. } => {
-                        let _ = reply.send(Vec::new());
-                    }
-                    _ => {}
-                }
-            }
-        });
-        (rib_tx, evpn_events_tx)
-    }
-
-    fn spawn_fake_peer_manager() -> (
-        mpsc::Sender<PeerManagerCommand>,
-        broadcast::Sender<SessionEvent>,
-        broadcast::Sender<PolicyEvent>,
-    ) {
-        let (peer_tx, mut peer_rx) = mpsc::channel(16);
-        let (session_events_tx, _) = broadcast::channel(16);
-        let (policy_events_tx, _) = broadcast::channel(16);
-        let session_events_tx_for_task = session_events_tx.clone();
-        let policy_events_tx_for_task = policy_events_tx.clone();
-        tokio::spawn(async move {
-            while let Some(command) = peer_rx.recv().await {
-                match command {
-                    PeerManagerCommand::SubscribeSessionEvents { reply } => {
-                        let _ = reply.send(session_events_tx_for_task.subscribe());
-                    }
-                    PeerManagerCommand::SubscribePolicyEvents { reply } => {
-                        let _ = reply.send(policy_events_tx_for_task.subscribe());
-                    }
-                    PeerManagerCommand::QueryPolicyEventHistory { reply, .. } => {
-                        let _ = reply.send(Vec::new());
-                    }
-                    _ => {}
-                }
-            }
-        });
-        (peer_tx, session_events_tx, policy_events_tx)
-    }
-
-    fn spawn_fake_peer_manager_with_history(
-        history: Vec<SessionLifecycleEvent>,
-    ) -> mpsc::Sender<PeerManagerCommand> {
-        let (peer_tx, mut peer_rx) = mpsc::channel(16);
-        let (events_tx, _) = broadcast::channel(16);
-        let history = Arc::new(history);
-        tokio::spawn(async move {
-            while let Some(command) = peer_rx.recv().await {
-                match command {
-                    PeerManagerCommand::SubscribeSessionEvents { reply } => {
-                        let _ = reply.send(events_tx.subscribe());
-                    }
-                    PeerManagerCommand::QuerySessionEventHistory {
-                        peer,
-                        event_types,
-                        limit,
-                        reply,
-                    } => {
-                        let limit = if limit == 0 {
-                            SESSION_EVENT_HISTORY_CAPACITY
-                        } else {
-                            limit.min(SESSION_EVENT_HISTORY_CAPACITY)
-                        };
-                        let mut events: Vec<_> = history
-                            .iter()
-                            .rev()
-                            .filter(|event| match peer {
-                                Some(peer) => event.peer == peer,
-                                None => true,
-                            })
-                            .filter(|event| {
-                                event_types.is_empty() || event_types.contains(&event.event_type)
-                            })
-                            .take(limit)
-                            .cloned()
-                            .collect();
-                        events.reverse();
-                        let _ = reply.send(events);
-                    }
-                    _ => {}
-                }
-            }
-        });
-        peer_tx
-    }
-
-    fn spawn_fake_peer_manager_with_policy_history(
-        history: Vec<PolicyEvent>,
-    ) -> mpsc::Sender<PeerManagerCommand> {
-        let (peer_tx, mut peer_rx) = mpsc::channel(16);
-        let (events_tx, _) = broadcast::channel(16);
-        let history = Arc::new(history);
-        tokio::spawn(async move {
-            while let Some(command) = peer_rx.recv().await {
-                match command {
-                    PeerManagerCommand::SubscribePolicyEvents { reply } => {
-                        let _ = reply.send(events_tx.subscribe());
-                    }
-                    PeerManagerCommand::QueryPolicyEventHistory { peer, limit, reply } => {
-                        let limit = if limit == 0 {
-                            POLICY_EVENT_HISTORY_CAPACITY
-                        } else {
-                            limit.min(POLICY_EVENT_HISTORY_CAPACITY)
-                        };
-                        let mut events: Vec<_> = history
-                            .iter()
-                            .rev()
-                            .filter(|event| match peer {
-                                Some(peer) => event.peer == Some(peer),
-                                None => true,
-                            })
-                            .take(limit)
-                            .cloned()
-                            .collect();
-                        events.reverse();
-                        let _ = reply.send(events);
-                    }
-                    _ => {}
-                }
-            }
-        });
-        peer_tx
-    }
-
-    fn session_event(peer: IpAddr, event_type: SessionLifecycleEventType) -> SessionEvent {
-        SessionEvent::Lifecycle(lifecycle_event(peer, event_type))
-    }
-
-    fn lifecycle_event(
-        peer: IpAddr,
-        event_type: SessionLifecycleEventType,
-    ) -> SessionLifecycleEvent {
-        SessionLifecycleEvent {
-            event_type,
-            peer,
-            peer_label: None,
-            timestamp: "456".to_string(),
-            old_state: Some(SessionState::OpenConfirm),
-            new_state: Some(SessionState::Established),
-            session_role: Some("primary".to_string()),
-            reason: format!("session event for peer {peer}"),
-        }
-    }
-
-    fn notification_event(peer: IpAddr, event_type: SessionNotificationEventType) -> SessionEvent {
-        SessionEvent::Notification(SessionNotificationEvent {
-            event_type,
-            peer,
-            timestamp: "789".to_string(),
-            code: 6,
-            subcode: 7,
-            description: "Connection Collision Resolution".to_string(),
-            session_role: Some("primary".to_string()),
-            shutdown_reason: None,
-            reason: format!(
-                "BGP NOTIFICATION sent for peer {peer}: 6/7 (Connection Collision Resolution)"
-            ),
-        })
-    }
-
-    fn policy_event(peer: Option<IpAddr>) -> PolicyEvent {
-        PolicyEvent {
-            operation: "set",
-            target_type: "policy",
-            target: "audit-policy".to_string(),
-            peer,
-            affected_peer_count: 2,
-            timestamp: "789".to_string(),
-            reason: "policy set policy audit-policy".to_string(),
-        }
-    }
-
-    fn fib_status(state: proto::FibRouteState) -> proto::FibRouteStatus {
-        proto::FibRouteStatus {
-            table_name: "blue".to_string(),
-            table_id: 100,
-            metric: 20,
-            prefix: "203.0.113.0".to_string(),
-            prefix_length: 24,
-            next_hop: "192.0.2.1".to_string(),
-            peer_address: "10.0.0.1".to_string(),
-            state: state as i32,
-            reason: "test".to_string(),
-            ..Default::default()
-        }
-    }
-
-    fn dataplane_route_event(
-        event_type: proto::BgpEventType,
-        prefix: &str,
-        prefix_length: u32,
-        peer_address: &str,
-    ) -> proto::BgpEvent {
-        let action = match event_type {
-            proto::BgpEventType::DataplaneRouteInstalled => "installed",
-            proto::BgpEventType::DataplaneRouteWithdrawn => "withdrawn",
-            proto::BgpEventType::DataplaneRouteFailed => "failed",
-            _ => "unknown",
-        };
-        proto::BgpEvent {
-            timestamp: "123".to_string(),
-            category: proto::EventCategory::Dataplane as i32,
-            event_type: event_type as i32,
-            severity: proto::EventSeverity::Info as i32,
-            peer_address: peer_address.to_string(),
-            previous_peer_address: String::new(),
-            prefix: prefix.to_string(),
-            prefix_length,
-            afi_safi: if prefix.contains(':') {
-                proto::AddressFamily::Ipv6Unicast as i32
-            } else {
-                proto::AddressFamily::Ipv4Unicast as i32
-            },
-            summary: format!("dataplane fib route {action} {prefix}/{prefix_length}"),
-            target_peer_address: String::new(),
-            event_id: None,
-            payload: Some(proto::bgp_event::Payload::DataplaneRoute(
-                proto::DataplaneRouteEvent {
-                    source: "fib".to_string(),
-                    action: action.to_string(),
-                    table_name: "blue".to_string(),
-                    table_id: 100,
-                    metric: 20,
-                    prefix: prefix.to_string(),
-                    prefix_length,
-                    next_hop: "192.0.2.1".to_string(),
-                    peer_address: peer_address.to_string(),
-                    timestamp: "123".to_string(),
-                    reason: action.to_string(),
-                },
-            )),
-        }
-    }
-
-    fn blackhole_status(state: proto::BlackholeDiscardState) -> proto::BlackholeDiscard {
-        proto::BlackholeDiscard {
-            prefix: "198.51.100.1".to_string(),
-            prefix_length: 32,
-            peer_address: "10.0.0.2".to_string(),
-            state: state as i32,
-            reason: "test".to_string(),
-        }
-    }
 
     #[tokio::test]
     async fn route_event_bridge_emits_bgp_event() {
